fix: components 分解
